+++ R/partial_signature.R
@@ -11,7 +11,7 @@
 #' 
 #' Remove the contribution of a SuperSig from the data and return the data.
 #' 
-#' @param dt a data frame of mutations (see vignette for details)
+#' @param data a data frame of mutations (see vignette for details)
 #' @param object an object of class "SuperSig"
 #' 
 #' @import dplyr
@@ -26,25 +26,25 @@
 #' head(example_dt) # use example data from package
 #' input_dt <- make_matrix(example_dt) # convert to correct format
 #' input_dt$IndVar <- c(1, 1, 1, 0, 0) # add IndVar column
-#' supersig <- get_signature(dt = input_dt, factor = "age") # get SuperSig
-#' partial_signature(dt = input_dt, object = supersig)
+#' supersig <- get_signature(data = input_dt, factor = "age") # get SuperSig
+#' partial_signature(data = input_dt, object = supersig)
 #' 
-partial_signature <- function(dt, object){
-    # Check column names of dt
-    if(is.na(match("SAMPLE_ID",toupper(colnames(dt))))){
+partial_signature <- function(data, object){
+    # Check column names of data
+    if(is.na(match("SAMPLE_ID",toupper(colnames(data))))){
         stop('Input data frame missing sample_id column.')
     } else {
-        colnames(dt)[match("SAMPLE_ID",toupper(colnames(dt)))] = "sample_id"
+        colnames(data)[match("SAMPLE_ID",toupper(colnames(data)))] = "sample_id"
     }
-    if(is.na(match("AGE",toupper(colnames(dt))))){
+    if(is.na(match("AGE",toupper(colnames(data))))){
         stop('Input data frame missing AGE column.')
     } else {
-        colnames(dt)[match("AGE",toupper(colnames(dt)))] = "AGE"
+        colnames(data)[match("AGE",toupper(colnames(data)))] = "AGE"
     }
-    if(is.na(match("INDVAR",toupper(colnames(dt))))){
+    if(is.na(match("INDVAR",toupper(colnames(data))))){
         stop('Input data frame missing IndVar column.')
     } else {
-        colnames(dt)[match("INDVAR",toupper(colnames(dt)))] = "IndVar"
+        colnames(data)[match("INDVAR",toupper(colnames(data)))] = "IndVar"
     }
     
     # Extract slots from object
@@ -54,10 +54,10 @@ partial_signature <- function(dt, object){
     for(i in seq_along(features)){
         trinucleotide_features <- unname(features[[i]])
         diff <- Signature(object)[[i]]
-        dt <- dt %>%
+        data <- data %>%
             mutate_at(vars(all_of(trinucleotide_features)), ~.- diff*AGE) %>%
             mutate_at(vars(all_of(trinucleotide_features)), ~ifelse(. < 0, 0, .))
     }
     
-    return(dt)
+    return(data)
 }