+++ R/predict_signature.R
@@ -28,10 +28,12 @@
 #' head(example_dt) # use example data from package
 #' input_dt <- make_matrix(example_dt) # convert to correct format
 #' input_dt$IndVar <- c(1, 1, 1, 0, 0) # add IndVar column
-#' out <- get_signature(dt = input_dt, factor = "Age") # get SuperSig
+#' head(input_dt)
+#' out <- get_signature(data = input_dt, factor = "Age") # get SuperSig
 #' 
 #' newdata <- predict_signature(out, newdata = input_dt, factor = "age")
 #' suppressPackageStartupMessages({library(dplyr)})
+#' head(newdata)
 #' head(newdata %>% select(X1, score))
 #' 
 predict_signature <- function(object, newdata, factor){